refactor: Async 적용되지 않던 문제 해결
